requirements.txt
langchain
langchain-groq
langchain-openai
feedparser
requests
beautifulsoup4
python-dotenv
tqdm
arxiv
